feat(api): add task creation and detail view support
Introduce a new API module for tasks and update the board columns and sample data to support task creation and detail views. This enables users to create tasks through a dialog and view task details in a dedicated view, improving the overall task management workflow.

Auto-committed-on: dragonfly
